requirements.txt
PyQt6
zeroconf
noiseprotocol